+++ bluetooth_low_energy_platform_interface/lib/src/advertise_data.dart
@@ -3,13 +3,10 @@ import 'dart:typed_data';
 import 'manufacturer_specific_data.dart';
 import 'uuid.dart';
 
-/// The advertisement discovered from a peripheral.
-class Advertisement {
+/// The advertise data discovered from a peripheral.
+class AdvertiseData {
   /// The name of the peripheral.
   final String? name;
-
-  /// The manufacturer specific data of the peripheral.
-  final ManufacturerSpecificData? manufacturerSpecificData;
 
   /// The GATT service uuids of the peripheral.
   final List<UUID> serviceUUIDs;
@@ -17,11 +14,14 @@ class Advertisement {
   /// The GATT service data of the peripheral.
   final Map<UUID, Uint8List> serviceData;
 
-  /// Constructs an [Advertisement].
-  Advertisement({
+  /// The manufacturer specific data of the peripheral.
+  final ManufacturerSpecificData? manufacturerSpecificData;
+
+  /// Constructs an [AdvertiseData].
+  AdvertiseData({
     this.name,
-    this.manufacturerSpecificData,
     this.serviceUUIDs = const [],
     this.serviceData = const {},
+    this.manufacturerSpecificData,
   });
 }